This reverts part of commit b82e48d83663afead2b5b0ed4c5f2fb2b32683ad.
And fix V type in ::set.

+++ clang/include/clang/Basic/Sanitizers.h
@@ -160,7 +160,7 @@ class SanitizerMaskCutoffs {
 public:
   std::optional<double> operator[](unsigned Kind) const;
 
-  void set(SanitizerMask K, float V);
+  void set(SanitizerMask K, double V);
   void clear(SanitizerMask K = SanitizerKind::All);
 };
